requirements.txt
# Core Framework
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0

# AI/LLM Integration
anthropic==0.18.1
openai==1.12.0

# ServiceNow Integration
pysnow==0.7.17
requests==2.31.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Transcript Parsing
webvtt-py==0.4.6
pysrt==1.1.2

# CLI
click==8.1.7
rich==13.7.0

# Email Notifications
sendgrid==6.11.0

# Utilities
pyyaml==6.0.1
jinja2==3.1.3
pydantic==2.6.1

# Development
pytest==8.0.0
black==24.1.1
flake8==7.0.0
